Various Formatting, Organizational, & Logic Fixes
- Made EngineROM_Init return a bool instead of void
- Made EngineROM_Verify return a bool instead of a uint32_t
- Made #defines for various game.dat file properties
- Fixed annoying int comparison warning in mage_dialog_control.cpp
- Realized my test code wasn't successfully writing to the ROM, need to fix.
- Made the SD card just print debug info instead of ENGINE_PANICing when an SD card is not found
- Added debug print statement when no SD card is found during startup code sequence in util_sd_error()

+++ Software/src/games/mage/mage_dialog_control.cpp
@@ -116,7 +116,7 @@ void MageDialogControl::load(
 
 void MageDialogControl::loadNextScreen() {
 	currentMessageIndex = 0;
-	if(currentScreenIndex >= currentDialogScreenCount) {
+	if((uint32_t)currentScreenIndex >= currentDialogScreenCount) {
 		isOpen = false;
 		return;
 	}
